refactor(layout): update components structure
