+++ components/Layout/Footer.js
@@ -1,5 +1,5 @@
 import React from "react";
-import footerStyles from "../styles/Footer.module.css";
+import footerStyles from "../../styles/Footer.module.css";
 
 const Footer = () => {
   return (
