+++ components/Layout/Nav.js
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import navStyles from "../styles/Nav.module.css";
+import navStyles from "../../styles/Nav.module.css";
 
 const Nav = () => {
   return (
