+++ components/Layout/index.js
@@ -1,7 +1,7 @@
 import Nav from "./Nav";
 import Meta from "./Meta";
 import Footer from "./Footer";
-import styles from "../styles/Layout.module.css";
+import styles from "../../styles/Layout.module.css";
 
 const Layout = ({ children }) => {
   return (
